@@ -4,17 +4,17 @@ description: >-
   that is used for communicating with Movesense devices. This plugin was createad by 
   Tugberk Agdokan and it used to be at https://github.com/tugberka/mdsflutter.
   
-version: 1.0.1
+version: 1.1.0-dev.1
 homepage: https://github.com/petri-lipponen-movesense/mdsflutter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  protobuf: ^1.1.4
+  protobuf: ^2.1.0
 
 dev_dependencies:
   flutter_test:
