@@ -9,7 +9,7 @@ class Mds {
   /// devices will get UUID as the address parameter. Scanning is terminated
   /// automatically after 60 seconds. Only devices with Movesense services are
   /// returned.
-  static void startScan(void Function(String, String) onNewDeviceFound) {
+  static void startScan(void Function(String?, String?) onNewDeviceFound) {
     MdsImpl().startScan(onNewDeviceFound);
   }
 
